==> src/2_EDA_inegi.R <==
### EDA ###

# Librerias ----
library(tidyverse)
library(readr)

# Lectura ----
tradicional <- read_csv("./data/2_interim/delitos_tradicional.csv")
control <- read_csv("./data/2_interim/delitos_control.csv")
oral <- read_csv("./data/2_interim/delitos_oral.csv")

# Graficos ----

## Tradicional ----
fig_tradicional <- tradicional %>%
  filter(anio != "2016") %>%
  group_by(concepto, anio) %>%
  summarise(freq = n()) %>%
  mutate(anio = as.character(anio)) %>%
  ggplot(aes(anio, freq, color = as.factor(concepto), label = freq)) +
  geom_point(size = 3, position = position_dodge(width = 0.8)) +
  geom_linerange(aes(x = anio, ymin = 0, ymax = freq, colour = as.factor(concepto)), 
                 position = position_dodge(width = 0.8)) +
  geom_text(vjust = -1, 
            position = position_dodge(width = 0.8)) +
  scale_color_discrete(name = "Tipo de delito") +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10), limits = c(0, 500)) +
  labs(title = "Número de delitos registrados por año en el municipio de Aguascalientes",
       subtitle = glue::glue("Sistema tradicional\n Total de delitos: {nrow(tradicional %>% filter(anio != '2016'))}"),
       x = "Año",
       y = "Cantidad de delitos",
       caption = "Elaboración propia con datos de INEGI (Impartición de justicia en materia penal).") +
  theme_light() +
  theme(title = element_text(size = 14),
        axis.text.x = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.text.y = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 13),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank()
  )

ggsave(plot = fig_tradicional, "./figs/delitos_tradicional.png", width = 10, height = 6, units = "in", dpi = 600)

## Control ----
fig_control <- control %>%
  group_by(concepto, anio) %>%
  summarise(freq = n()) %>%
  mutate(anio = as.character(anio)) %>%
  ggplot(aes(anio, freq, color = as.factor(concepto), label = freq)) +
  geom_point(size = 4, position = position_dodge(width = 0.8)) +
  geom_linerange(aes(x = anio, ymin = 0, ymax = freq, colour = as.factor(concepto)), 
                 position = position_dodge(width = 0.8)) +
  geom_text(vjust = -1, 
            position = position_dodge(width = 0.8)) +
  scale_color_discrete(name = "Tipo de delito") +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10), limits = c(0, 500)) +
  labs(title = "Número de delitos registrados por año en el municipio de Aguascalientes",
       subtitle = glue::glue("Sistema Acusatorio Oral. Juzgado de Control\n Total de delitos: {nrow(control)}"),
       x = "Año",
       y = "Cantidad de delitos",
       caption = "Elaboración propia con datos de INEGI (Impartición de justicia en materia penal).") +
  theme_light() +
  theme(title = element_text(size = 14),
        axis.text.x = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.text.y = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 13),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank()
  )

ggsave(plot = fig_control, "./figs/delitos_control.png", width = 10, height = 6, units = "in", dpi = 600)

## Oral ----
fig_oral <- oral %>%
  group_by(concepto, anio) %>%
  summarise(freq = n()) %>%
  mutate(anio = as.character(anio)) %>%
  ggplot(aes(anio, freq, color = as.factor(concepto), label = freq)) +
  geom_point(size = 3, position = position_dodge(width = 0.8)) +
  geom_linerange(aes(x = anio, ymin = 0, ymax = freq, colour = as.factor(concepto)), 
                 position = position_dodge(width = 0.8)) +
  geom_text(vjust = -1, 
            position = position_dodge(width = 0.8)) +
  scale_color_discrete(name = "Tipo de delito") +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10), limits = c(0, 50)) +
  labs(title = "Número de delitos registrados por año en el municipio de Aguascalientes",
       subtitle = glue::glue("Sistema Acusatorio Oral. Juzgado de Juicio Oral\n Total de delitos: {nrow(oral)}"),
       x = "Año",
       y = "Cantidad de delitos",
       caption = "Elaboración propia con datos de INEGI (Impartición de justicia en materia penal).") +
  theme_light() +
  theme(title = element_text(size = 14),
        axis.text.x = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.text.y = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 13),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank()
  )

ggsave(plot = fig_oral, "./figs/delitos_oral.png", width = 10, height = 6, units = "in", dpi = 600)
